# libs/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barbearia Premium API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  /** @nullable */
  imageUrl: string | null;
}

export interface TimeSlot {
  /** Time in HH:MM format */
  time: string;
  available: boolean;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  serviceId: number;
  serviceName: string;
  servicePrice: number;
  /** Date in YYYY-MM-DD format */
  appointmentDate: string;
  /** Time in HH:MM format */
  appointmentTime: string;
  status: AppointmentStatus;
  /** @nullable */
  paymentId: string | null;
  createdAt: string;
}

export interface CreateCheckoutBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  serviceId: number;
  /** Date in YYYY-MM-DD format */
  appointmentDate: string;
  /** Time in HH:MM format */
  appointmentTime: string;
}

export interface CheckoutSession {
  sessionId: string;
  amount: number;
  serviceName: string;
  appointmentDate: string;
  appointmentTime: string;
  customerName: string;
}

export interface ConfirmAppointmentBody {
  sessionId: string;
  paymentMethodId: string;
}

export interface StatsSummary {
  totalAppointments: number;
  todayAppointments: number;
  totalRevenue: number;
  /** @nullable */
  popularService: string | null;
}

export type GetAvailabilityParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
  serviceId: number;
};