# libs/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barbearia Premium API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /availability:
    get:
      operationId: getAvailability
      tags: [availability]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /appointments/confirm:
    post:
      operationId: confirmAppointment
      tags: [appointments]
      summary: Confirm appointment after successful payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmAppointmentBody"
      responses:
        "201":
          description: Appointment confirmed and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Bad request or invalid payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Time slot no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TimeSlot:
      type: object
      properties:
        time:
          type: string
          description: Time in HH:MM format
        available:
          type: boolean
      required:
        - time
        - available
    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        servicePrice:
          type: number
        appointmentDate:
          type: string
          description: Date in YYYY-MM-DD format
        appointmentTime:
          type: string
          description: Time in HH:MM format
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        paymentId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - serviceId
        - serviceName
        - servicePrice
        - appointmentDate
        - appointmentTime
        - status
        - paymentId
        - createdAt
...
    ConfirmAppointmentBody:
      type: object
      properties:
        sessionId:
          type: string
        paymentMethodId:
          type: string
      required:
        - sessionId
        - paymentMethodId
...