# libs/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barbearia Premium API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all available services
 */
export const ListServicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  durationMinutes: zod.number(),
  imageUrl: zod.string().nullable(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * @summary Get available time slots for a date
 */
export const GetAvailabilityQueryParams = zod.object({
  date: zod.coerce.string(),
  serviceId: zod.coerce.number(),
});

export const GetAvailabilityResponseItem = zod.object({
  time: zod.string().describe("Time in HH:MM format"),
  available: zod.boolean(),
});
export const GetAvailabilityResponse = zod.array(GetAvailabilityResponseItem);

/**
 * @summary List all confirmed appointments
 */
export const ListAppointmentsResponseItem = zod.object({
  id: zod.number(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
  serviceId: zod.number(),
  serviceName: zod.string(),
  servicePrice: zod.number(),
  appointmentDate: zod.string().describe("Date in YYYY-MM-DD format"),
  appointmentTime: zod.string().describe("Time in HH:MM format"),
  status: zod.enum(["pending", "confirmed", "cancelled"]),
  paymentId: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListAppointmentsResponse = zod.array(ListAppointmentsResponseItem);

/**
 * @summary Create a payment session for an appointment
 */
export const CreateCheckoutBody = zod.object({
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
  serviceId: zod.number(),
  appointmentDate: zod.string().describe("Date in YYYY-MM-DD format"),
  appointmentTime: zod.string().describe("Time in HH:MM format"),
});

export const CreateCheckoutResponse = zod.object({
  sessionId: zod.string(),
  amount: zod.number(),
  serviceName: zod.string(),
  appointmentDate: zod.string(),
  appointmentTime: zod.string(),
  customerName: zod.string(),
});

/**
 * @summary Confirm appointment after successful payment
 */
export const ConfirmAppointmentBody = zod.object({
  sessionId: zod.string(),
  paymentMethodId: zod.string(),
});

/**
 * @summary Get appointment by ID
 */
export const GetAppointmentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAppointmentResponse = zod.object({
  id: zod.number(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
  serviceId: zod.number(),
  serviceName: zod.string(),
  servicePrice: zod.number(),
  appointmentDate: zod.string().describe("Date in YYYY-MM-DD format"),
  appointmentTime: zod.string().describe("Time in HH:MM format"),
  status: zod.enum(["pending", "confirmed", "cancelled"]),
  paymentId: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Get barbershop stats summary
 */
export const GetStatsSummaryResponse = zod.object({
  totalAppointments: zod.number(),
  todayAppointments: zod.number(),
  totalRevenue: zod.number(),
  popularService: zod.string().nullable(),
});