# libs/api-zod/src/generated/types/confirmAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barbearia Premium API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmAppointmentBody {
  sessionId: string;
  paymentMethodId: string;
}